Add new `CycleBuilder` method
